+++ blocks/webcam.h
@@ -26,7 +26,7 @@
 /* ../blocks/webcam.c */
 
 char *
-b_write_webcam_on(char *dst, unsigned int dst_size, const char *unused, unsigned int *interval);
+b_write_webcam_on(char *dst, unsigned int dst_size, const char *unused, unsigned short *interval);
 
 #	endif
